pyproject.toml
[tool.poetry]
name = "pptx2aituberkitscript"
version = "0.1.0"
description = "This package converts pptx to markdown"
repository = "https://github.com/JunzoKamahara/pptx2aituberkitscript"
authors = ["Junzo Kamahara <j.kamahara@gmail.com>"]
license = "MIT Licence"

[tool.poetry.scripts]
pptx2md = "pptx2aituberkitscript.__main__:main"

[tool.poetry.dependencies]
python = ">=3.7"
python-pptx = ">=0.6.18"
rapidfuzz = ">=0.10.0"
Pillow = ">=6.0.0"
tqdm = ">=4"

[tool.poetry.dev-dependencies]
yapf = ">=0.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
column_limit = 120
